pyproject.toml
[tool.poetry]
name = "good-toolbelt"
version = "0.1.0"
description = ""
authors = ["Chris Goddard <chris@goodkiwi.llc>"]
readme = "README.md"
packages = [{include = "good_toolbelt"}]

[tool.poetry.dependencies]
python = "^3.11"
pyyaml = "^6.0.1"
ipykernel = "^6.24.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
